Rename package `std.sap` to `sapnwrfc`.
This was a suggestion on the forum. This library will probably not
become part of the D standard library. Therefore the `std.sap` prefix
is no good choice. The new name `sapnwrfc` matches library names from
other languages, e.g. Python, Ruby. Hopefully this makes it easier to
discover this library.